2.1pre20
{
    "filename": "links-2.1pre20.tar.gz",
    "md5": "4c360c208bb7105e868668445644b5cc",
    "sha1": "a73f7528fd0e47f49451bd8b6638802e045277f0",
    "sha256": "4e92e6207e3020c2dd6c143ac9282d585ce593bfb38178cb9068b9b2007779bd"
}

+++ main.c
@@ -172,7 +172,7 @@ int attach_g_terminal(void *info, int len)
 #endif
 
 struct object_request *dump_obj;
-int dump_pos;
+off_t dump_pos;
 
 void end_dump(struct object_request *r, void *p)
 {
